rm_watermark/requirements.txt
Pillow==10.2.0
numpy==1.26.4
scipy==1.12.0
opencv-python==4.9.0.80 